Move and integrate all unit tests in one folder (preparation for Travis CI)
