+++ .travis.yml
@@ -1,0 +1,26 @@
+language: python
+python:
+    - 2.6
+    - 2.7
+    - 3.2
+    - 3.3
+    - 3.4
+    - pypy
+    - pypy3
+
+install:
+ - pip install --install-option='--no-cython-compile' cython
+
+script:
+ - python setup.py build_ext --inplace
+ - nosetests -vv
+
+notifications:
+    email:
+        on_success: change
+        on_failure: change
+
+matrix:
+  allow_failures:
+    - python: pypy
+    - python: pypy3
