+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include LICENSE.txt
+include README.rst
