--- alltests.py
@@ -1,7 +1,0 @@
-import unittest
-
-from rstest import *
-from polynomialtest import *
-from fftest import *
-
-unittest.main()
