+++ tests/test_rs.py
